requirements.txt
PyQt5==5.15.10
psutil==5.9.8
PyYAML==6.0.2
paho-mqtt==2.1.0
apscheduler==3.10.4
qtawesome==1.3.1
amqtt==0.11.0
pyqtgraph==0.13.3
